Remove superfluous properties from TileOjbect.
TileObject's layer_id, tile_id, format_id, and gridset_id are
properties of type long that are not used at all throughout the
whole code base. They're leftovers from the removal of the
"metastore" which was an embedded SQL database that assigned
these identifiers to each tile's row in the metadata table.