Fix Azure blobstore error instantiating AzureClient
commit 72ce566b "Avoid infinite number of calls to listBlobs when
doing prefix removals (e..g, gridset or layer removals)"
changed the gwc `AzureClient` check for container existence from an
attemp to create it to an attempt to get its properties before creating
it.

When the container doesn't exist, the blocking call throws
`com.microsoft.azure.storage.blob.StorageException`, which contains the
status code.

Wrap that call in a try-catch block and get the status code from the
exception.