Run Azure blob store integration tests with Azure and Azurite
Azurite is run as a testcontainer, and the github CI build runs all
integration tests for Java 11, 17, and 21.

In order to save on Azure resources, online integration tests are
run only if the Azurite based tests succeeded, using github secrets
to set up the account, key, and container; and for Java 11 only.